# Add frozen string pragma