--- a/benchmarks/filters.rb
+++ b/benchmarks/filters.rb
@@ -21,10 +21,7 @@ VALUES = {
   date:      [Date.new, Date.new.to_s],
   date_time: [DateTime.new, DateTime.new.to_s],
   decimal:   [BigDecimal.new(0)],
-  file:      [
-    File.new(__FILE__),
-    Struct.new(:tempfile).new(File.new(__FILE__))
-  ],
+  file:      [File.new(__FILE__)],
   float:     [0.0, '0.0', 0],
   hash:      [{}], # TODO
   integer:   [0, '0', 0.0],