--- a/benchmarks/filters.rb
+++ b/benchmarks/filters.rb
@@ -1,56 +1,34 @@
 # coding: utf-8
 
-# Usage: bundle exec ruby benchmarks/filters.rb
-
-# TODO: Benchmark errors.
-# TODO: Special cases for values and filters (see other TODOs).
-# TODO: Parse (and graph) the output.
-# TODO: Automate benchmarks across time (i.e., commits).
-
 require 'active_interaction'
-require 'active_support/inflector'
-require 'benchmark'
-
-filters = ActiveInteraction::Filter
-  .send(:const_get, :CLASSES)
-  .reject { |_, v| v.name =~ /\bAbstract/ }
+require 'benchmark/ips'
 
 VALUES = {
-  array:     [[]], # TODO
-  boolean:   [false, '0', 'false', true, '1', 'true'],
-  date:      [Date.new, Date.new.to_s],
-  date_time: [DateTime.new, DateTime.new.to_s],
-  decimal:   [BigDecimal.new(0)],
-  file:      [File.new(__FILE__)],
-  float:     [0.0, '0.0', 0],
-  hash:      [{}], # TODO
-  integer:   [0, '0', 0.0],
-  interface: [Object.new],
-  object:    [Object.new], # TODO: Reconstantizing.
-  string:    [''], # TODO: Without strip.
-  symbol:    [:'', ''],
-  time:      [Time.at(0), Time.at(0).to_s, 0] # TODO: TimeWithZone
+  array: [],
+  boolean: false,
+  date: Date.new,
+  date_time: DateTime.new,
+  decimal: BigDecimal.new(0),
+  file: StringIO.new,
+  float: 0.0,
+  hash: {},
+  integer: 0,
+  interface: Object.new,
+  object: Object.new,
+  string: '',
+  symbol: :'',
+  time: Time.at(0)
 }
 
-missing_filters = filters.keys - VALUES.keys
-fail "Missing filters: #{missing_filters}" unless missing_filters.empty?
+Benchmark.ips do |bm|
+  bm.report('lambda') { -> {}.call }
 
-Benchmark.bmbm do |bm|
-  n = Integer(ARGV.pop || 10_000)
-
-  bm.report('baseline') do
-    n.times { 100.times.reduce('') { |a, e| a + (e % 255).chr } }
-  end
-
-  VALUES.each do |slug, values|
-    filter = ActiveInteraction::Filter.factory(slug).new(slug, class: Object)
-
-    values.each do |value|
-      bm.report("#{slug}.cast(#{value.inspect})") do
-        n.times do
-          filter.cast(value)
-        end
-      end
+  VALUES.each do |type, value|
+    interaction = Class.new(ActiveInteraction::Base) do
+      def execute; end
     end
+    interaction.public_send(type, type)
+
+    bm.report(type) { interaction.run!(type => value) }
   end
 end