--- a/lib/active_interaction/backports.rb
+++ b/lib/active_interaction/backports.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 require 'active_support/core_ext'