--- a/lib/active_interaction/base.rb
+++ b/lib/active_interaction/base.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 require 'active_support/core_ext/hash/indifferent_access'
 require 'set'