--- a/lib/active_interaction/concerns/active_modelable.rb
+++ b/lib/active_interaction/concerns/active_modelable.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 module ActiveInteraction
   # Implement the minimal ActiveModel interface.