--- a/lib/active_interaction/concerns/hashable.rb
+++ b/lib/active_interaction/concerns/hashable.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 module ActiveInteraction
   # Allow `hash` to be overridden when given arguments and/or a block.