--- a/lib/active_interaction/concerns/missable.rb
+++ b/lib/active_interaction/concerns/missable.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 module ActiveInteraction
   # Handle common `method_missing` functionality.