--- a/lib/active_interaction/filter.rb
+++ b/lib/active_interaction/filter.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 require 'active_support/inflector'