--- a/lib/active_interaction/filter_column.rb
+++ b/lib/active_interaction/filter_column.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 module ActiveInteraction
   # A minimal implementation of an `ActiveRecord::ConnectionAdapters::Column`.