--- a/lib/active_interaction/filters/string_filter.rb
+++ b/lib/active_interaction/filters/string_filter.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 module ActiveInteraction
   class Base