--- a/lib/active_interaction/grouped_input.rb
+++ b/lib/active_interaction/grouped_input.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 require 'ostruct'