--- a/lib/active_interaction/modules/input_processor.rb
+++ b/lib/active_interaction/modules/input_processor.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 module ActiveInteraction
   # Groups inputs ending in "(*N*i)" into {GroupedInput}.