--- a/lib/active_interaction/modules/validation.rb
+++ b/lib/active_interaction/modules/validation.rb
@@ -1,4 +1,5 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 module ActiveInteraction
   # Validates inputs using filters.