Fixes #335. 

Tested locally with Ruby 2.3.0-preview2. All of the tests pass.