@tfausak 2015-12-17
Raw benchmark results.

# Without pragma

```
Rehearsal ----------------------------------------------------------------------------------------
baseline                                               0.350000   0.000000   0.350000 (  0.347545)
array.cast([])                                         0.070000   0.000000   0.070000 (  0.070843)
boolean.cast(false)                                    0.000000   0.000000   0.000000 (  0.000876)
boolean.cast("0")                                      0.000000   0.000000   0.000000 (  0.001399)
boolean.cast("false")                                  0.010000   0.000000   0.010000 (  0.008608)
boolean.cast(true)                                     0.010000   0.000000   0.010000 (  0.003816)
boolean.cast("1")                                      0.000000   0.000000   0.000000 (  0.005089)
boolean.cast("true")                                   0.010000   0.000000   0.010000 (  0.011282)
date.cast(Mon, 01 Jan -4712)                           0.010000   0.000000   0.010000 (  0.003477)
date.cast("-4712-01-01")                               0.040000   0.000000   0.040000 (  0.045156)
date_time.cast(Mon, 01 Jan -4712 00:00:00 +0000)       0.010000   0.000000   0.010000 (  0.004416)
date_time.cast("-4712-01-01T00:00:00+00:00")           0.130000   0.000000   0.130000 (  0.136035)
decimal.cast(#<BigDecimal:7fb80b8c9928,'0.0',9(27)>)   0.010000   0.000000   0.010000 (  0.005341)
file.cast(#<File:benchmarks/filters.rb>)               0.000000   0.000000   0.000000 (  0.005937)
float.cast(0.0)                                        0.010000   0.000000   0.010000 (  0.001328)
float.cast("0.0")                                      0.000000   0.000000   0.000000 (  0.006389)
float.cast(0)                                          0.010000   0.000000   0.010000 (  0.005686)
hash.cast({})                                          0.020000   0.000000   0.020000 (  0.018301)
integer.cast(0)                                        0.000000   0.000000   0.000000 (  0.001392)
integer.cast("0")                                      0.000000   0.000000   0.000000 (  0.006313)
integer.cast(0.0)                                      0.010000   0.000000   0.010000 (  0.005507)
interface.cast(#<Object:0x007fb80b8c96f8>)             0.000000   0.000000   0.000000 (  0.002544)
object.cast(#<Object:0x007fb80b8c96a8>)                0.000000   0.000000   0.000000 (  0.001537)
string.cast("")                                        0.000000   0.000000   0.000000 (  0.001919)
symbol.cast(:"")                                       0.010000   0.000000   0.010000 (  0.000882)
symbol.cast("")                                        0.000000   0.000000   0.000000 (  0.001748)
time.cast(1969-12-31 18:00:00 -0600)                   0.020000   0.000000   0.020000 (  0.019735)
time.cast("1969-12-31 18:00:00 -0600")                 0.220000   0.000000   0.220000 (  0.222317)
time.cast(0)                                           0.010000   0.000000   0.010000 (  0.009298)
------------------------------------------------------------------------------- total: 0.960000sec

                                                           user     system      total        real
baseline                                               0.330000   0.000000   0.330000 (  0.332318)
array.cast([])                                         0.060000   0.000000   0.060000 (  0.061277)
boolean.cast(false)                                    0.000000   0.000000   0.000000 (  0.000891)
boolean.cast("0")                                      0.000000   0.000000   0.000000 (  0.001329)
boolean.cast("false")                                  0.010000   0.000000   0.010000 (  0.007105)
boolean.cast(true)                                     0.000000   0.000000   0.000000 (  0.002696)
boolean.cast("1")                                      0.000000   0.000000   0.000000 (  0.004531)
boolean.cast("true")                                   0.010000   0.000000   0.010000 (  0.009972)
date.cast(Mon, 01 Jan -4712)                           0.000000   0.000000   0.000000 (  0.002646)
date.cast("-4712-01-01")                               0.040000   0.000000   0.040000 (  0.037857)
date_time.cast(Mon, 01 Jan -4712 00:00:00 +0000)       0.000000   0.000000   0.000000 (  0.002632)
date_time.cast("-4712-01-01T00:00:00+00:00")           0.140000   0.000000   0.140000 (  0.136277)
decimal.cast(#<BigDecimal:7fb80b8c9928,'0.0',9(27)>)   0.000000   0.000000   0.000000 (  0.004083)
file.cast(#<File:benchmarks/filters.rb>)               0.000000   0.000000   0.000000 (  0.003151)
float.cast(0.0)                                        0.000000   0.000000   0.000000 (  0.001351)
float.cast("0.0")                                      0.000000   0.000000   0.000000 (  0.004568)
float.cast(0)                                          0.000000   0.000000   0.000000 (  0.003968)
hash.cast({})                                          0.010000   0.000000   0.010000 (  0.014121)
integer.cast(0)                                        0.000000   0.000000   0.000000 (  0.001316)
integer.cast("0")                                      0.010000   0.000000   0.010000 (  0.005010)
integer.cast(0.0)                                      0.010000   0.000000   0.010000 (  0.004449)
interface.cast(#<Object:0x007fb80b8c96f8>)             0.000000   0.000000   0.000000 (  0.002119)
object.cast(#<Object:0x007fb80b8c96a8>)                0.000000   0.000000   0.000000 (  0.001528)
string.cast("")                                        0.000000   0.000000   0.000000 (  0.001793)
symbol.cast(:"")                                       0.000000   0.000000   0.000000 (  0.000915)
symbol.cast("")                                        0.000000   0.000000   0.000000 (  0.001778)
time.cast(1969-12-31 18:00:00 -0600)                   0.010000   0.000000   0.010000 (  0.010914)
time.cast("1969-12-31 18:00:00 -0600")                 0.210000   0.000000   0.210000 (  0.216155)
time.cast(0)                                           0.010000   0.000000   0.010000 (  0.006659)
```

# With pragma

```
Rehearsal ----------------------------------------------------------------------------------------
baseline                                               0.340000   0.000000   0.340000 (  0.346002)
array.cast([])                                         0.070000   0.000000   0.070000 (  0.067100)
boolean.cast(false)                                    0.000000   0.000000   0.000000 (  0.000882)
boolean.cast("0")                                      0.000000   0.000000   0.000000 (  0.001298)
boolean.cast("false")                                  0.010000   0.000000   0.010000 (  0.008528)
boolean.cast(true)                                     0.000000   0.000000   0.000000 (  0.003169)
boolean.cast("1")                                      0.010000   0.000000   0.010000 (  0.004988)
boolean.cast("true")                                   0.010000   0.000000   0.010000 (  0.011551)
date.cast(Mon, 01 Jan -4712)                           0.000000   0.000000   0.000000 (  0.003519)
date.cast("-4712-01-01")                               0.050000   0.000000   0.050000 (  0.046200)
date_time.cast(Mon, 01 Jan -4712 00:00:00 +0000)       0.000000   0.000000   0.000000 (  0.004371)
date_time.cast("-4712-01-01T00:00:00+00:00")           0.150000   0.000000   0.150000 (  0.143409)
decimal.cast(#<BigDecimal:7fd6f2fd0630,'0.0',9(27)>)   0.000000   0.000000   0.000000 (  0.004994)
file.cast(#<File:benchmarks/filters.rb>)               0.010000   0.000000   0.010000 (  0.005719)
float.cast(0.0)                                        0.000000   0.000000   0.000000 (  0.001325)
float.cast("0.0")                                      0.000000   0.000000   0.000000 (  0.005404)
float.cast(0)                                          0.010000   0.000000   0.010000 (  0.004752)
hash.cast({})                                          0.020000   0.000000   0.020000 (  0.017149)
integer.cast(0)                                        0.000000   0.000000   0.000000 (  0.001336)
integer.cast("0")                                      0.000000   0.000000   0.000000 (  0.005799)
integer.cast(0.0)                                      0.010000   0.000000   0.010000 (  0.005256)
interface.cast(#<Object:0x007fd6f2fd0360>)             0.000000   0.000000   0.000000 (  0.002465)
object.cast(#<Object:0x007fd6f2fd0310>)                0.000000   0.000000   0.000000 (  0.001532)
string.cast("")                                        0.000000   0.000000   0.000000 (  0.001875)
symbol.cast(:"")                                       0.000000   0.000000   0.000000 (  0.000872)
symbol.cast("")                                        0.010000   0.000000   0.010000 (  0.001741)
time.cast(1969-12-31 18:00:00 -0600)                   0.010000   0.000000   0.010000 (  0.017472)
time.cast("1969-12-31 18:00:00 -0600")                 0.220000   0.000000   0.220000 (  0.219048)
time.cast(0)                                           0.010000   0.000000   0.010000 (  0.009174)
------------------------------------------------------------------------------- total: 0.940000sec

                                                           user     system      total        real
baseline                                               0.330000   0.000000   0.330000 (  0.328084)
array.cast([])                                         0.060000   0.000000   0.060000 (  0.059688)
boolean.cast(false)                                    0.000000   0.000000   0.000000 (  0.000944)
boolean.cast("0")                                      0.000000   0.000000   0.000000 (  0.001299)
boolean.cast("false")                                  0.010000   0.000000   0.010000 (  0.007424)
boolean.cast(true)                                     0.000000   0.000000   0.000000 (  0.002653)
boolean.cast("1")                                      0.000000   0.000000   0.000000 (  0.004463)
boolean.cast("true")                                   0.010000   0.000000   0.010000 (  0.010398)
date.cast(Mon, 01 Jan -4712)                           0.000000   0.000000   0.000000 (  0.002642)
date.cast("-4712-01-01")                               0.030000   0.000000   0.030000 (  0.038681)
date_time.cast(Mon, 01 Jan -4712 00:00:00 +0000)       0.000000   0.000000   0.000000 (  0.002627)
date_time.cast("-4712-01-01T00:00:00+00:00")           0.120000   0.000000   0.120000 (  0.126175)
decimal.cast(#<BigDecimal:7fd6f2fd0630,'0.0',9(27)>)   0.010000   0.000000   0.010000 (  0.004213)
file.cast(#<File:benchmarks/filters.rb>)               0.010000   0.000000   0.010000 (  0.003023)
float.cast(0.0)                                        0.000000   0.000000   0.000000 (  0.001312)
float.cast("0.0")                                      0.000000   0.000000   0.000000 (  0.004752)
float.cast(0)                                          0.000000   0.000000   0.000000 (  0.004010)
hash.cast({})                                          0.010000   0.000000   0.010000 (  0.015030)
integer.cast(0)                                        0.000000   0.000000   0.000000 (  0.001321)
integer.cast("0")                                      0.000000   0.000000   0.000000 (  0.004835)
integer.cast(0.0)                                      0.000000   0.000000   0.000000 (  0.004457)
interface.cast(#<Object:0x007fd6f2fd0360>)             0.000000   0.000000   0.000000 (  0.002089)
object.cast(#<Object:0x007fd6f2fd0310>)                0.000000   0.000000   0.000000 (  0.001517)
string.cast("")                                        0.000000   0.000000   0.000000 (  0.001784)
symbol.cast(:"")                                       0.000000   0.000000   0.000000 (  0.000985)
symbol.cast("")                                        0.010000   0.000000   0.010000 (  0.001779)
time.cast(1969-12-31 18:00:00 -0600)                   0.010000   0.000000   0.010000 (  0.010996)
time.cast("1969-12-31 18:00:00 -0600")                 0.210000   0.000000   0.210000 (  0.215601)
time.cast(0)                                           0.010000   0.000000   0.010000 (  0.006602)
```