@tfausak 2015-12-17
With the new benchmark, most things are faster by about 2%.

![untitled](https://cloud.githubusercontent.com/assets/408753/11877531/7eb73d14-a4b5-11e5-846e-9399b82a8ee8.png)

thawed | frozen | delta
--- | --- | ---
24.392 | 24.828 | 0.436
44.547 | 44.698 | 0.151
42.281 | 42.968 | 0.687
42.545 | 43.067 | 0.522
40.988 | 41.874 | 0.886
42.507 | 42.571 | 0.064
43.666 | 44.398 | 0.732
34.002 | 34.798 | 0.796
43.789 | 44.160 | 0.371
43.114 | 42.774 | -0.340
43.504 | 44.270 | 0.766
43.444 | 43.700 | 0.256
44.320 | 44.749 | 0.429
36.616 | 37.079 | 0.463

# Without pragma

```
              lambda      2.066M (± 5.7%) i/s -     10.324M
               array     24.392k (± 2.6%) i/s -    123.180k
             boolean     44.547k (± 1.9%) i/s -    223.320k
                date     42.281k (± 2.4%) i/s -    213.256k
           date_time     42.545k (± 2.4%) i/s -    214.140k
             decimal     40.988k (± 2.7%) i/s -    206.382k
                file     42.507k (± 2.1%) i/s -    215.100k
               float     43.666k (± 2.1%) i/s -    221.220k
                hash     34.002k (± 1.8%) i/s -    171.390k
             integer     43.789k (± 2.2%) i/s -    220.620k
           interface     43.114k (± 2.2%) i/s -    216.977k
              object     43.504k (± 2.0%) i/s -    217.474k
              string     43.444k (± 2.2%) i/s -    219.720k
              symbol     44.320k (± 2.5%) i/s -    222.253k
                time     36.616k (± 2.8%) i/s -    184.440k
```

# With pragma

```
              lambda      2.072M (± 4.8%) i/s -     10.334M
               array     24.828k (± 1.7%) i/s -    124.410k
             boolean     44.698k (± 2.1%) i/s -    223.474k
                date     42.968k (± 1.9%) i/s -    216.630k
           date_time     43.067k (± 1.7%) i/s -    218.772k
             decimal     41.874k (± 2.0%) i/s -    211.294k
                file     42.571k (± 1.6%) i/s -    216.398k
               float     44.398k (± 2.8%) i/s -    223.996k
                hash     34.798k (± 1.6%) i/s -    175.674k
             integer     44.160k (± 2.2%) i/s -    221.966k
           interface     42.774k (± 3.1%) i/s -    214.704k
              object     44.270k (± 2.2%) i/s -    221.730k
              string     43.700k (± 2.3%) i/s -    220.516k
              symbol     44.749k (± 2.2%) i/s -    223.725k
                time     37.079k (± 2.0%) i/s -    187.746k
```